Improve compatibility of the config across platforms
The init script should now work better across different platforms
(specifically Windows and Linux, at the time of this writing),
installing only the packages needed on the specific platform the
configuration is being loaded on.
In other words, no package will be installed but not used in the init
script.