Switch to locally-maintained package list
I figured out that it's hard to keep a file that is automatically
written by Emacs consistent with the different packages that I want to
install on Linux and Windows. I decided to go for a custom list
maintained by me in one of my init files and then write that list after
processing to a custom.el file, which *isn't* checked in.