Improve company behavior when scrolling through selection
company will no longer complete the common part of a selection when
trying to scroll through the candidates. I initially did that because
I thought that was the only way to get company to wrap around the
candidates when reaching either the top or the bottom of the list,
then I found out about the company-selection-wrap-around variable.

Further, selecting the previous/next candidate is now remapped from
previous-line and next-line respectively instead of hardcoded to C-p
and C-n.